Moving  classmethods from metaclasses to actual class definition. It gets much cleaner now.
