+++ prophy/composite.py
@@ -1,5 +1,5 @@
 from . import scalar
-from .data_types import base_array, distance_to_next_multiply, struct_packed
+from .data_types import base_array, build_container_length_field, distance_to_next_multiply, struct_packed
 from .desc_item import descriptor_item_type, codec_kind
 from .exception import ProphyError
 from .six import long, string_types
@@ -55,42 +55,6 @@ class _composite_generator_base(_generator_base):
 
     def add_sizers(self):
         pass
-
-    def _bricks_walk(self, cursor):
-        def eval_path(leaf_path):
-            return ".%s%s" % (item.name, leaf_path or "")
-
-        for item in self._descriptor:
-
-            padding_size = cursor.distance_to_next(item.type._ALIGNMENT)
-            if padding_size:
-                yield make_padding(padding_size), eval_path(".:pre_padding")
-
-            for sub_brick_type, sub_brick_path in item.type._bricks_walk(cursor):
-                yield sub_brick_type, eval_path(sub_brick_path)
-
-            if item.type._PARTIAL_ALIGNMENT:
-                padding_size = cursor.distance_to_next(item.type._PARTIAL_ALIGNMENT)
-                if padding_size:
-                    yield make_padding(padding_size), eval_path(".:partial_padding")
-
-        padding_size = cursor.distance_to_next(self._ALIGNMENT)
-        if padding_size:
-            yield make_padding(padding_size), eval_path(".:final_padding")
-
-    def get_descriptor(self):
-        """
-            FIXME: I'm afraid it rapes YAGNI rule
-        """
-        return [item.descriptor_info for item in self._descriptor]
-
-    def wire_pattern(self):
-        cursor = _cursor_class()
-        for type_, path_ in self._bricks_walk(cursor):
-            type_size = getattr(type_, "_SIZE", "??")
-            if type_size != "??":
-                cursor.pos += type_size
-            yield path_, type_.__name__, type_size
 
 
 class enum_generator(_generator_base):
@@ -169,19 +133,20 @@ class struct_generator(_composite_generator_base):
                 alignment = 1
             alignment = max(type_._ALIGNMENT, alignment)
         if not issubclass(self, struct_packed) and self._descriptor:
-            self._SIZE += sum(self.get_padded_sizes())
 
-    def get_padded_sizes(self):
-        types = list(self._types())
-        sizes = [tp._SIZE for tp in types]
-        alignments = [tp._ALIGNMENT for tp in types[1:]] + [self._ALIGNMENT]
-        offset = 0
+            def get_padded_sizes():
+                types = list(self._types())
+                sizes = [tp._SIZE for tp in types]
+                alignments = [tp._ALIGNMENT for tp in types[1:]] + [self._ALIGNMENT]
+                offset = 0
 
-        for size, alignment in zip(sizes, alignments):
-            offset += size
-            padding = distance_to_next_multiply(offset, alignment)
-            offset += padding
-            yield padding
+                for size, alignment in zip(sizes, alignments):
+                    offset += size
+                    padding = distance_to_next_multiply(offset, alignment)
+                    offset += padding
+                    yield padding
+
+            self._SIZE += sum(get_padded_sizes())
 
     def add_properties(self):
         for item in self._descriptor:
@@ -394,72 +359,3 @@ class union_generator(_composite_generator_base):
             new_value = item.type._check(new_value)
             self_._fields[item.name] = new_value
         setattr(self, item.name, property(getter, setter))
-
-
-def build_container_length_field(sizer_item_type, container_name, bound_shift):
-    class container_len(sizer_item_type):
-        _BOUND = [container_name]
-
-        @classmethod
-        def add_bounded_container(cls, cont_name):
-            cls._BOUND.append(cont_name)
-
-        @classmethod
-        def evaluate_size(cls, parent):
-            sizes = set(len(getattr(parent, c_name)) for c_name in cls._BOUND)
-            if len(sizes) != 1:
-                msg = "Size mismatch of arrays in {}: {}"
-                raise ProphyError(msg.format(parent.__class__.__name__, ", ".join(cls._BOUND)))
-            return sizes.pop()
-
-        @staticmethod
-        def _encode(value, endianness):
-            return sizer_item_type._encode(value + bound_shift, endianness)
-
-        @staticmethod
-        def _decode(data, pos, endianness):
-            value, size = sizer_item_type._decode(data, pos, endianness)
-            array_guard = 65536
-            if value > array_guard:
-                raise ProphyError("decoded array length over %s" % array_guard)
-            value -= bound_shift
-            if value < 0:
-                raise ProphyError("decoded array length smaller than shift")
-            return value, size
-
-        @classmethod
-        def _bricks_walk(cls, _):
-            yield sizer_item_type, " (sizer)"
-
-    return container_len
-
-
-class _cursor_class(object):
-    """
-        A helper for breaking variables scope while passing the "self.pos" between _bricks_walk iterators call.
-        TODO: It will be probably not needed.
-    """
-
-    def __init__(self):
-        self.pos = 0
-
-    def distance_to_next(self, alignment):
-        remainer = self.pos % alignment
-        return (alignment - remainer) % alignment
-
-
-def make_padding(padding_size):
-    class PaddingMock(object):
-        _SIZE = padding_size
-
-        @staticmethod
-        def encode_mock():
-            return b'\x00' * padding_size
-
-        @staticmethod
-        def decode_mock(data, pos):
-            if (len(data) - pos) < padding_size:
-                raise ProphyError("too few bytes to decode padding")
-            return data[pos:(pos + padding_size)], padding_size
-    PaddingMock.__name__ = "<Pd{}>".format(padding_size)
-    return PaddingMock
